move Delete icon inside editor/
