+++ icons/editor/Delete.tsx
